feat(#2): add option to hide skipped tasks
Added a new configuration option `hide_skipped` that allows users to
hide tasks that were skipped in the Smart Commit UI. This makes the
task list cleaner when many conditional tasks are present.

The feature is enabled by default in the main config but can be
toggled per-project. Implemented the necessary UI filtering logic
to respect this setting.